# academy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Summer Academy educational platform",
        "title": "Summer Academy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/days": {
            "get": {
                "security": [
                    {
                        "JWTCookie": []
                    }
                ],
                "description": "Displays a list of all available days with their problems",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "problems"
                ],
                "summary": "List available days",
                "responses": {
                    "200": {
                        "description": "Days list page"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/leaderboard": {
            "get": {
                "description": "Displays a leaderboard with top users and their scores",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Show leaderboard page",
                "responses": {
                    "200": {
                        "description": "Leaderboard page"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/login": {
            "get": {
                "description": "Renders the login page for user authentication",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Show the login page",
                "responses": {
                    "200": {
                        "description": "Login page"
                    }
                }
            },
            "post": {
                "description": "Validates OTP, creates/updates user, and issues JWT token on successful login",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Process login form submission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OTP code",
                        "name": "otp",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to days page"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/verify": {
            "get": {
                "description": "Renders the OTP verification page with phone number and OTP if provided",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Show OTP verification page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "OTP code",
                        "name": "otp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification page"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "description": "Bearer JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "JWTCookie": {
            "description": "JWT token stored in cookie for authentication",
            "type": "apiKey",
            "name": "academy_session",
            "in": "cookie"
        }
    }
}